backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==2.3.7
PyJWT==2.8.0
bcrypt==4.0.1
Flask-Mail==0.9.1
Pillow==10.0.1
sendgrid==6.9.7
mega.py==1.0.8
